Expand TestLogger concept to allow testing error logging or standard logging
