+++ test/TestLog.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "TestLogger.h"
+
+class TestLogger;
+
+void ResetTestLog();
+
+TestLogger& GetLogger();
+TestLogger& GetErrorLogger();
